requirements.txt
torch
torchvision
torchaudio
hydra-core
tqdm
wandb
einops
tqdm
